Compare NONMEM runs estimation results

Returns a table summarizing the results of estimation steps of a list of NONMEM runs.

compare_runs(runs = NULL, estimation_number = NULL)

Arguments

runs
a list of pmxploit NONMEM run objects.

estimation_number
(optional) integer. A vector of containing the position(s) of the estimation step(s) to compare in the runs. If estimation_number=-1, the last estimation step of each run will be selected. Default is NULL to consider all estimation steps.

Value

A data frame.

Details

Results are gathered in a nested data frame, containing one row per estimation step included in the comparison. Each row contains the following columns:

RUN_ID: the run identifying name/number.

INFO: information about the run (problem, number of observations, number of individuals and run duration).

FILES: a tibble containing the control stream name, the dataset name and the run archive path.

ESTIMATION: a nested tibble containing the following columns:

SUMMARY: a tibble containing the estimation results summary (method, OFV, termination status, duration...).

THETA: a tibble containing THETA parameters results.

OMEGA: a tibble containing OMEGA parameters results.

SIGMA: a tibble containing SIGMA parameters results.

ETABAR: a tibble containing ETABAR results.

SHRINKAGE: a tibble containing shrinking results.

Examples

compare_runs(list(EXAMPLERUN, EXAMPLERUN))

#> # A tibble: 2 x 4
#>   RUN_ID         INFO             FILES            ESTIMATION      
#>   <chr>          <list>           <list>           <list>          
#> 1 TMDD_Djebli_al <tibble [1 × 5]> <tibble [1 × 3]> <tibble [1 × 6]>
#> 2 TMDD_Djebli_al <tibble [1 × 5]> <tibble [1 × 3]> <tibble [1 × 6]>