Load a NONMEM run data

Loads NONMEM run results data from either a folder or a *.tar.gz archive file.

load_nm_run(path, temp_directory = str_c(tempdir(), "pmxploit"),
  load_tables = TRUE, read_initial_values = TRUE,
  keep_tempfiles = FALSE, extract_everything = FALSE,
  dataset_separator = NA, dataset_na_string = ".",
  update_progress = NULL, verbose = FALSE)

Arguments

path
character. Run folder or archive file path.

temp_directory
(optional) character. When path is a *.tar.gz archive file, sets the path of the temporary directory where the archive files will be extracted.

load_tables
logical. If TRUE (default), loads output tables.

read_initial_values
logical. If TRUE (default), parses initial parameter values from the control stream.

keep_tempfiles
logical. If TRUE, temp_directory will not be deleted once run data is loaded.

extract_everything
logical. If TRUE, when path is a *.tar.gz archive file, extracts all the content of the archive. Otherwise, extracts only the files required for post-processing analysis (default).

dataset_separator
(optional) character. Character used as column separator in the dataset. Default is NA for automatic detection.

dataset_na_string
character. Character string corresponding to missing values in the dataset. Default is ".".

update_progress
(otional) function of the form function(detail, ...){}. Useful to follow loading progression.

verbose
logical. If TRUE: prints output messages.

Value

A NONMEM run object. See load_nm_run_directory.

Examples

# NOT RUN {
run <- load_nm_run("path/to/my/run_folder")

run <- load_nm_run("path/to/my/run_archive.tar.gz")
# }