Starts a NONMEM run

Starts a NONMEM run

nm_exec(control_file, run_directory = NULL, nonmem_exe = NULL,
  n_nodes = 1L, parafile = NULL, result_directory = NULL,
  job_id = NULL, archive = TRUE, cleanup = TRUE, quiet = FALSE,
  call_template = NULL, ...)

Arguments

control_file
path of the control file

run_directory
path of the directory where to perform the run

nonmem_exe
path of NONMEM executable path

n_nodes
number of CPUs

parafile
path of a NONMEM parallel processing settings file

result_directory
path of the directory where to place the results

job_id
(optional) integer run ID

archive
logical. If TRUE, compress the result files to a *.tar.gz archive?

cleanup
logical. If TRUE, remove NONMEM temporary files

quiet
logical. If TRUE, do not print NONMEM output when executing

call_template
character string used to construct NONMEM call. Helps creating custom script calls.

Value

A pmxploit NONMEM run object.

Examples

\notrun{
nm_exec(control_file = "control_file.ctl", nonmem_exe = "/path/to/nmfe")
}

#> Error: <text>:1:1: unexpected input
#> 1: \
#>     ^