Generate a NONMEM control stream with $PRIOR

Generates a NONMEM control stream for an analysis using prior information from a source run.

nm_prior(run, df_formula = 0, df_values = NULL, output_file = NULL,
  dataset_filename = NULL, estimations = NULL)

Arguments

run
pmxploit NONMEM run object of the source run.

df_formula
integer. Formula to use for computation of degrees of freedom of OMEGA and SIGMA priors.

output_file
(optional) character. File path to write the generated control stream to. Default is NULL, not saving any file.

dataset_filename
(optional) character. Name of the dataset to set in the $DATA record of the generated control stream. Default is NULL, leaving $DATA record as it is in the source run.

estimations
(optional) list. A list containing the estimation record(s) to set for the generated control stream. Each estimation record must be represented by a list item containing the estimation settings as a named list.

Value

A character string of the generated control stream code.

Examples

# Control stream generation from pmxploit::EXAMPLERUN
cs1 <- nm_prior(EXAMPLERUN)
# see the output
writeLines(cs1)

#> ; Control stream generated for PRIOR analysis from:
#> ; ~/pmxploit/example.tar.gz
#> 
#> $PROB  TMDD Qss model for alirocumab
#> ; Reference:
#> ; Djebli N, Martinez JM, Lohan L, Khier S, Brunet A, Hurbin F, et al.
#> ; Target-mediated drug disposition population pharmacokinetics model of alirocumab in
#> ; healthy volunteers and patients: pooled analysis of randomized phase I/II/III studies.
#> ; Clin Pharmacokinet. 2017
#> 
#> $INPUT ;; Edit $INPUT to match the new dataset columns
#> $DATA dataset.csv IGNORE=@
#> 
#> $SUBROUTINE ADVAN13 TOL=5
#> 
#> $PRIOR NWPRI
#> 
#> $MODEL NCOMP=4
#> COMP=(DEPOT)
#> COMP=(SARTOT)
#> COMP=(TPCSK9)
#> COMP=(PERIPH)
#> 
#> $PK
#> 
#> TVCL = THETA(1)
#> TKON = THETA(2)
#> TKIN = THETA(3)
#> TKDE = THETA(4)
#> TVQ  = THETA(5)
#> TVV1 = THETA(6)
#> TVV2 = THETA(7)
#> TVKA = THETA(8)
#> TVF  = THETA(9)
#> TVEP = THETA(10)
#> TVEA = THETA(11)
#> TVLAG = THETA(12)
#> COV1   = THETA(13)
#> 
#> 
#> ETCL   = ETA(1)
#> EKON   = ETA(2)
#> EKIN   = ETA(3)
#> EKDE   = ETA(4)
#> ETQ    = ETA(5)
#> ETV1   = ETA(6)
#> ETV2   = ETA(7)
#> ETKA   = ETA(8)
#> ETF    = ETA(9)
#> 
#> 
#> 
#> CL   = TVCL * EXP(ETCL)
#> KON  = TKON * EXP(EKON)
#> KINT = TKIN * EXP(EKIN)
#> KDEG = TKDE * EXP(EKDE)
#> Q    = TVQ  * EXP(ETQ)
#> 
#> 
#> 
#> COV1V1 = TVV1 * COV1 ** DISST
#> 
#> V1 = COV1V1 * EXP(ETV1)
#> 
#> V2   = TVV2 * EXP(ETV2)
#> KA   = TVKA * EXP(ETKA)
#> ALAG1 = TVLAG
#> 
#> IF (ISC .EQ.0) THEN
#> 
#> F1=1
#> ELSE
#> PHI = LOG(TVF/(1-TVF))
#> F1 = EXP(PHI+ETF)/(1+EXP(PHI+ETF))
#> ENDIF
#> 
#> KEL = CL / V1
#> KPT = Q / V1
#> KTP = Q / V2
#> 
#> KSS = 0.58 + KINT/KON
#> 
#> SARBASE = 0
#> KSYN = TBSPCSK * KDEG
#> 
#> A_0(2) = SARBASE
#> A_0(3) = TBSPCSK * V1
#> 
#> S2=V1
#> S3=V1
#> 
#> 
#> 
#> $DES
#> C2=A(2)/V1
#> C3=A(3)/V1
#> C4=A(4)/V2
#> DADT(1)= -KA * A(1) * ISC * F1
#> 
#> K1 = C2 - C3 - KSS
#> SARF = 0.5*(K1+ SQRT(K1*K1 + 4*KSS*C2))
#> 
#> DADT(2)= KA*A(1)*ISC*F1-(KEL+KPT)*SARF*V1-KINT*A(3)*SARF/(KSS+SARF)+KTP*C4*V2
#> 
#> DADT(3)= KSYN  * V1 - KDEG * A(3)- (KINT - KDEG) * SARF * A(3)/(KSS + SARF)
#> 
#> DADT(4)= KPT* SARF * V1 - KTP * A(4)
#> 
#> ;A(1) = DOSE
#> ;SARF = [SAR_LIBRE]
#> ;A(2) = [SARtot]*V1
#> ;A(3) = [PCSK9tot]*V1
#> ;A(4) = [L tissulaire]*V2
#> 
#> ; PCSK9SAR = TPCSK9 * SARF/(KSS + SARF) => surprediction possible du complexe au
#> ;tps initial !!!
#> 
#> $ERROR
#> 
#> IPRED=F
#> SAV=0
#> IF (IPRED.EQ.0) SAV=0.00001
#> W=F
#> IRES=DV-IPRED
#> IWRES=IRES/(W+SAV)
#> 
#> Y=F+SQRT(TVEA*TVEA + TVEP*TVEP*F*F)*EPS(1)
#> 
#> $THETA ( 0.0 0.20 1000000.0 ) ; TVCL
#> $THETA ( 559 559 559 ) ; TKON
#> $THETA ( 0.0 0.25 1000000.0 ) ; TKIN
#> $THETA ( 0.0 2.00 1000000.0 ) ; TKDE
#> $THETA ( 0.0 1.00 1000000.0 ) ; TVQ
#> $THETA ( 2.0 4.00 1000000.0 ) ; TVV1 SART
#> $THETA (   2.612   2.612       2.612 ) ; TVV2
#> $THETA ( 0.0 0.50 1000000.0 ) ; TVKA
#> $THETA (   0.550   0.750       1.000 ) ; TVF
#> $THETA ( 0.0 0.30 1000000.0 ) ; TVEP
#> $THETA ( 0.0 0.05 1000000.0 ) ; TVEA
#> $THETA ( 0.0 0.05  100) ; LAG
#> $THETA 1 ; COV1
#> 
#> 
#> $OMEGA 0.5
#> $OMEGA 0 FIX
#> $OMEGA 0.5
#> $OMEGA 0.5
#> $OMEGA 0.5
#> $OMEGA 0.5
#> $OMEGA 0 FIX
#> $OMEGA 0.5
#> $OMEGA 0.5
#> 
#> $SIGMA  1 FIX
#> 
#> 
#> ; ================================================
#> ; PRIOR INFORMATION
#> ; ================================================
#> ; THETA Priors -------------------------
#> $THETAP
#> 0.171896943763098     FIXED
#> 559                   FIXED
#> 0.123803201342408     FIXED
#> 1.34635526853289      FIXED
#> 0.50039445791587      FIXED
#> 3.22590460107421      FIXED
#> 2.612                 FIXED
#> 0.637779367431747     FIXED
#> 0.590001808531246     FIXED
#> 0.274684075637132     FIXED
#> 0.581795606608002     FIXED
#> 0.0298266165438046    FIXED
#> 1.56428570804632      FIXED
#> ; Variance to THETA priors -------------
#> $THETAPV BLOCK(13)
#> 0.00023104719622956   FIXED
#> 1e-09                 1e-09                
#> -1.04034957452488e-05 1e-09                 4.94852473541275e-06 
#> -0.000132380576300711 1e-09                 -2.17038124549764e-06 0.0014067543164689   
#> 3.37273335283364e-05  1e-09                 -2.62117718892531e-05 -9.83509242784927e-05 0.00081696660001029  
#> -0.000710872133743716 1e-09                 -5.80489374980969e-05 -0.000732628454663643 -0.000445920587930254 0.0415711705221499   
#> 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                
#> -0.000172472594375445 1e-09                 1.49084134799218e-05  0.000179969508697223  -0.000100076823054098 -0.00224596953903649  1e-09                 0.00218763549223713  
#> 8.79808997968756e-05  1e-09                 -2.09503076577778e-05 -0.000106886489009519 4.53146822800296e-05  0.00309692392362989   1e-09                 -0.000478502787321484 0.000572988006168684 
#> -1.16269709887021e-06 1e-09                 6.45639736622959e-07  -1.04783992387998e-05 -4.49284209190946e-06 8.68593698728809e-06  1e-09                 2.1408518826492e-06   -2.72270313463236e-06 2.26505265649685e-06 
#> 3.97881008538303e-06  1e-09                 4.80556912014364e-07  2.78111595286765e-05  -3.57590444667545e-06 -4.77538174170191e-05 1e-09                 -6.15715850827622e-05 2.27771577491074e-06  -9.41191843254687e-06 0.000500919762819647 
#> 5.45053324304558e-07  1e-09                 -1.7273759122248e-07  -1.94999218108576e-07 3.31278452211055e-06  9.25019594253438e-07  1e-09                 -7.15492047000474e-07 5.74018896478916e-08  -1.6453046982605e-07  5.38469288515014e-06  1.20876103617239e-06 
#> 0.000507900274801002  1e-09                 -5.23098619372009e-05 -0.00029555530917814  0.000667518750332241  -0.0099825252948127   1e-09                 -0.000951749317422468 0.000240924919937593  -2.13162869496016e-05 3.10211676261024e-05  1.05759826160997e-06  0.00723573763017683  
#> ; OMEGA Priors -------------------------
#> $OMEGAP
#> 0.279195982109823     FIXED
#> 0                     FIXED
#> 0.0580908084573451    FIXED
#> 0.170534475475969     FIXED
#> 0.0691065862986685    FIXED
#> 0.0990369682744904    FIXED
#> 0                     FIXED
#> 0.464462143306044     FIXED
#> 0.289551101707683     FIXED
#> ; Degrees of freedom of OMEGA Prior ----
#> $OMEGAPD
#> 27                    FIXED ; ETCL
#> 0                     FIXED ; EKON
#> 316                   FIXED ; EKIN
#> 216                   FIXED ; EKDE
#> 11                    FIXED ; ETQ
#> 71                    FIXED ; ETV1
#> 0                     FIXED ; ETV2
#> 176                   FIXED ; ETKA
#> 33                    FIXED ; ETF
#> ; SIGMA Priors -------------------------
#> $SIGMAP
#> 1                     FIXED
#> ; Degrees of freedom of SIGMA Prior ----
#> $SIGMAPD
#> 0                     FIXED ; EPS1
#> ; ======================================
#> $ESTIMATION METHOD=1 INTERACTION PRINT=1 MAXEVAL=9999 NOABORT POSTHOC
#> 
#> $COVARIANCE PRINT=E MATRIX=S
#> 
#> $TAB ID TIME DV IPRED PRED CMT IRES CWRES IWRES NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=sdtab
#> $TAB ID TIME CL KON KSS KINT KDEG Q V1 V2 KA ALAG1 F1 ETAS(1:LAST) NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=patab
#> $TAB ID TIME AGE WT BMI CLCR GFR BSLDLC TBSPCSK FBSPCSK NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=cotab
#> $TAB ID TIME ISC DISST SEX FIBR EZE STATIN ARM ABDO THIGH STUD NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=catab
#> 

# Use prior information with "new_dataset.csv" and save the control stream to "new_study.ctl"
nm_prior(EXAMPLERUN, output_file = "new_study.ctl", dataset_filename = "new_dataset.csv")

#> [1] "; Control stream generated for PRIOR analysis from:\n; ~/pmxploit/example.tar.gz\n\n$PROB  TMDD Qss model for alirocumab\n; Reference:\n; Djebli N, Martinez JM, Lohan L, Khier S, Brunet A, Hurbin F, et al.\n; Target-mediated drug disposition population pharmacokinetics model of alirocumab in\n; healthy volunteers and patients: pooled analysis of randomized phase I/II/III studies.\n; Clin Pharmacokinet. 2017\n\n$INPUT ;; Edit $INPUT to match the new dataset columns\n$DATA new_dataset.csv IGNORE=@\n\n$SUBROUTINE ADVAN13 TOL=5\n\n$PRIOR NWPRI\n\n$MODEL NCOMP=4\nCOMP=(DEPOT)\nCOMP=(SARTOT)\nCOMP=(TPCSK9)\nCOMP=(PERIPH)\n\n$PK\n\nTVCL = THETA(1)\nTKON = THETA(2)\nTKIN = THETA(3)\nTKDE = THETA(4)\nTVQ  = THETA(5)\nTVV1 = THETA(6)\nTVV2 = THETA(7)\nTVKA = THETA(8)\nTVF  = THETA(9)\nTVEP = THETA(10)\nTVEA = THETA(11)\nTVLAG = THETA(12)\nCOV1   = THETA(13)\n\n\nETCL   = ETA(1)\nEKON   = ETA(2)\nEKIN   = ETA(3)\nEKDE   = ETA(4)\nETQ    = ETA(5)\nETV1   = ETA(6)\nETV2   = ETA(7)\nETKA   = ETA(8)\nETF    = ETA(9)\n\n\n\nCL   = TVCL * EXP(ETCL)\nKON  = TKON * EXP(EKON)\nKINT = TKIN * EXP(EKIN)\nKDEG = TKDE * EXP(EKDE)\nQ    = TVQ  * EXP(ETQ)\n\n\n\nCOV1V1 = TVV1 * COV1 ** DISST\n\nV1 = COV1V1 * EXP(ETV1)\n\nV2   = TVV2 * EXP(ETV2)\nKA   = TVKA * EXP(ETKA)\nALAG1 = TVLAG\n\nIF (ISC .EQ.0) THEN\n\nF1=1\nELSE\nPHI = LOG(TVF/(1-TVF))\nF1 = EXP(PHI+ETF)/(1+EXP(PHI+ETF))\nENDIF\n\nKEL = CL / V1\nKPT = Q / V1\nKTP = Q / V2\n\nKSS = 0.58 + KINT/KON\n\nSARBASE = 0\nKSYN = TBSPCSK * KDEG\n\nA_0(2) = SARBASE\nA_0(3) = TBSPCSK * V1\n\nS2=V1\nS3=V1\n\n\n\n$DES\nC2=A(2)/V1\nC3=A(3)/V1\nC4=A(4)/V2\nDADT(1)= -KA * A(1) * ISC * F1\n\nK1 = C2 - C3 - KSS\nSARF = 0.5*(K1+ SQRT(K1*K1 + 4*KSS*C2))\n\nDADT(2)= KA*A(1)*ISC*F1-(KEL+KPT)*SARF*V1-KINT*A(3)*SARF/(KSS+SARF)+KTP*C4*V2\n\nDADT(3)= KSYN  * V1 - KDEG * A(3)- (KINT - KDEG) * SARF * A(3)/(KSS + SARF)\n\nDADT(4)= KPT* SARF * V1 - KTP * A(4)\n\n;A(1) = DOSE\n;SARF = [SAR_LIBRE]\n;A(2) = [SARtot]*V1\n;A(3) = [PCSK9tot]*V1\n;A(4) = [L tissulaire]*V2\n\n; PCSK9SAR = TPCSK9 * SARF/(KSS + SARF) => surprediction possible du complexe au\n;tps initial !!!\n\n$ERROR\n\nIPRED=F\nSAV=0\nIF (IPRED.EQ.0) SAV=0.00001\nW=F\nIRES=DV-IPRED\nIWRES=IRES/(W+SAV)\n\nY=F+SQRT(TVEA*TVEA + TVEP*TVEP*F*F)*EPS(1)\n\n$THETA ( 0.0 0.20 1000000.0 ) ; TVCL\n$THETA ( 559 559 559 ) ; TKON\n$THETA ( 0.0 0.25 1000000.0 ) ; TKIN\n$THETA ( 0.0 2.00 1000000.0 ) ; TKDE\n$THETA ( 0.0 1.00 1000000.0 ) ; TVQ\n$THETA ( 2.0 4.00 1000000.0 ) ; TVV1 SART\n$THETA (   2.612   2.612       2.612 ) ; TVV2\n$THETA ( 0.0 0.50 1000000.0 ) ; TVKA\n$THETA (   0.550   0.750       1.000 ) ; TVF\n$THETA ( 0.0 0.30 1000000.0 ) ; TVEP\n$THETA ( 0.0 0.05 1000000.0 ) ; TVEA\n$THETA ( 0.0 0.05  100) ; LAG\n$THETA 1 ; COV1\n\n\n$OMEGA 0.5\n$OMEGA 0 FIX\n$OMEGA 0.5\n$OMEGA 0.5\n$OMEGA 0.5\n$OMEGA 0.5\n$OMEGA 0 FIX\n$OMEGA 0.5\n$OMEGA 0.5\n\n$SIGMA  1 FIX\n\n\n; ================================================\n; PRIOR INFORMATION\n; ================================================\n; THETA Priors -------------------------\n$THETAP\n0.171896943763098     FIXED\n559                   FIXED\n0.123803201342408     FIXED\n1.34635526853289      FIXED\n0.50039445791587      FIXED\n3.22590460107421      FIXED\n2.612                 FIXED\n0.637779367431747     FIXED\n0.590001808531246     FIXED\n0.274684075637132     FIXED\n0.581795606608002     FIXED\n0.0298266165438046    FIXED\n1.56428570804632      FIXED\n; Variance to THETA priors -------------\n$THETAPV BLOCK(13)\n0.00023104719622956   FIXED\n1e-09                 1e-09                \n-1.04034957452488e-05 1e-09                 4.94852473541275e-06 \n-0.000132380576300711 1e-09                 -2.17038124549764e-06 0.0014067543164689   \n3.37273335283364e-05  1e-09                 -2.62117718892531e-05 -9.83509242784927e-05 0.00081696660001029  \n-0.000710872133743716 1e-09                 -5.80489374980969e-05 -0.000732628454663643 -0.000445920587930254 0.0415711705221499   \n1e-09                 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                \n-0.000172472594375445 1e-09                 1.49084134799218e-05  0.000179969508697223  -0.000100076823054098 -0.00224596953903649  1e-09                 0.00218763549223713  \n8.79808997968756e-05  1e-09                 -2.09503076577778e-05 -0.000106886489009519 4.53146822800296e-05  0.00309692392362989   1e-09                 -0.000478502787321484 0.000572988006168684 \n-1.16269709887021e-06 1e-09                 6.45639736622959e-07  -1.04783992387998e-05 -4.49284209190946e-06 8.68593698728809e-06  1e-09                 2.1408518826492e-06   -2.72270313463236e-06 2.26505265649685e-06 \n3.97881008538303e-06  1e-09                 4.80556912014364e-07  2.78111595286765e-05  -3.57590444667545e-06 -4.77538174170191e-05 1e-09                 -6.15715850827622e-05 2.27771577491074e-06  -9.41191843254687e-06 0.000500919762819647 \n5.45053324304558e-07  1e-09                 -1.7273759122248e-07  -1.94999218108576e-07 3.31278452211055e-06  9.25019594253438e-07  1e-09                 -7.15492047000474e-07 5.74018896478916e-08  -1.6453046982605e-07  5.38469288515014e-06  1.20876103617239e-06 \n0.000507900274801002  1e-09                 -5.23098619372009e-05 -0.00029555530917814  0.000667518750332241  -0.0099825252948127   1e-09                 -0.000951749317422468 0.000240924919937593  -2.13162869496016e-05 3.10211676261024e-05  1.05759826160997e-06  0.00723573763017683  \n; OMEGA Priors -------------------------\n$OMEGAP\n0.279195982109823     FIXED\n0                     FIXED\n0.0580908084573451    FIXED\n0.170534475475969     FIXED\n0.0691065862986685    FIXED\n0.0990369682744904    FIXED\n0                     FIXED\n0.464462143306044     FIXED\n0.289551101707683     FIXED\n; Degrees of freedom of OMEGA Prior ----\n$OMEGAPD\n27                    FIXED ; ETCL\n0                     FIXED ; EKON\n316                   FIXED ; EKIN\n216                   FIXED ; EKDE\n11                    FIXED ; ETQ\n71                    FIXED ; ETV1\n0                     FIXED ; ETV2\n176                   FIXED ; ETKA\n33                    FIXED ; ETF\n; SIGMA Priors -------------------------\n$SIGMAP\n1                     FIXED\n; Degrees of freedom of SIGMA Prior ----\n$SIGMAPD\n0                     FIXED ; EPS1\n; ======================================\n$ESTIMATION METHOD=1 INTERACTION PRINT=1 MAXEVAL=9999 NOABORT POSTHOC\n\n$COVARIANCE PRINT=E MATRIX=S\n\n$TAB ID TIME DV IPRED PRED CMT IRES CWRES IWRES NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=sdtab\n$TAB ID TIME CL KON KSS KINT KDEG Q V1 V2 KA ALAG1 F1 ETAS(1:LAST) NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=patab\n$TAB ID TIME AGE WT BMI CLCR GFR BSLDLC TBSPCSK FBSPCSK NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=cotab\n$TAB ID TIME ISC DISST SEX FIBR EZE STATIN ARM ABDO THIGH STUD NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=catab\n"

# Use prior information with "new_dataset.csv", set estimation to SAEM/IMP and
# save the control stream to "new_study.ctl"
nm_prior(EXAMPLERUN,
         output_file = "new_study.ctl",
         dataset_filename = "new_dataset.csv",
         estimations = list(list(method = "SAEM",
                                 "INTERACTION", NBURN = 400, NITER = 1000,
                                 PRINT = 50, "NOABORT", CTYPE = 3),
                            list(method = "IMP",
                                 NITER = 10, PRINT = 1,
                                 ISAMPLE = 3000, EONLY = 1)))

#> [1] "$PROB  TMDD Qss model for alirocumab\n; Reference:\n; Djebli N, Martinez JM, Lohan L, Khier S, Brunet A, Hurbin F, et al.\n; Target-mediated drug disposition population pharmacokinetics model of alirocumab in\n; healthy volunteers and patients: pooled analysis of randomized phase I/II/III studies.\n; Clin Pharmacokinet. 2017\n\n$INPUT ;; Edit $INPUT to match the new dataset columns\n$DATA new_dataset.csv IGNORE=@\n\n$SUBROUTINE ADVAN13 TOL=5\n\n$PRIOR NWPRI\n\n$MODEL NCOMP=4\nCOMP=(DEPOT)\nCOMP=(SARTOT)\nCOMP=(TPCSK9)\nCOMP=(PERIPH)\n\n$PK\n\nTVCL = THETA(1)\nTKON = THETA(2)\nTKIN = THETA(3)\nTKDE = THETA(4)\nTVQ  = THETA(5)\nTVV1 = THETA(6)\nTVV2 = THETA(7)\nTVKA = THETA(8)\nTVF  = THETA(9)\nTVEP = THETA(10)\nTVEA = THETA(11)\nTVLAG = THETA(12)\nCOV1   = THETA(13)\n\n\nETCL   = ETA(1)\nEKON   = ETA(2)\nEKIN   = ETA(3)\nEKDE   = ETA(4)\nETQ    = ETA(5)\nETV1   = ETA(6)\nETV2   = ETA(7)\nETKA   = ETA(8)\nETF    = ETA(9)\n\n\n\nCL   = TVCL * EXP(ETCL)\nKON  = TKON * EXP(EKON)\nKINT = TKIN * EXP(EKIN)\nKDEG = TKDE * EXP(EKDE)\nQ    = TVQ  * EXP(ETQ)\n\n\n\nCOV1V1 = TVV1 * COV1 ** DISST\n\nV1 = COV1V1 * EXP(ETV1)\n\nV2   = TVV2 * EXP(ETV2)\nKA   = TVKA * EXP(ETKA)\nALAG1 = TVLAG\n\nIF (ISC .EQ.0) THEN\n\nF1=1\nELSE\nPHI = LOG(TVF/(1-TVF))\nF1 = EXP(PHI+ETF)/(1+EXP(PHI+ETF))\nENDIF\n\nKEL = CL / V1\nKPT = Q / V1\nKTP = Q / V2\n\nKSS = 0.58 + KINT/KON\n\nSARBASE = 0\nKSYN = TBSPCSK * KDEG\n\nA_0(2) = SARBASE\nA_0(3) = TBSPCSK * V1\n\nS2=V1\nS3=V1\n\n\n\n$DES\nC2=A(2)/V1\nC3=A(3)/V1\nC4=A(4)/V2\nDADT(1)= -KA * A(1) * ISC * F1\n\nK1 = C2 - C3 - KSS\nSARF = 0.5*(K1+ SQRT(K1*K1 + 4*KSS*C2))\n\nDADT(2)= KA*A(1)*ISC*F1-(KEL+KPT)*SARF*V1-KINT*A(3)*SARF/(KSS+SARF)+KTP*C4*V2\n\nDADT(3)= KSYN  * V1 - KDEG * A(3)- (KINT - KDEG) * SARF * A(3)/(KSS + SARF)\n\nDADT(4)= KPT* SARF * V1 - KTP * A(4)\n\n;A(1) = DOSE\n;SARF = [SAR_LIBRE]\n;A(2) = [SARtot]*V1\n;A(3) = [PCSK9tot]*V1\n;A(4) = [L tissulaire]*V2\n\n; PCSK9SAR = TPCSK9 * SARF/(KSS + SARF) => surprediction possible du complexe au\n;tps initial !!!\n\n$ERROR\n\nIPRED=F\nSAV=0\nIF (IPRED.EQ.0) SAV=0.00001\nW=F\nIRES=DV-IPRED\nIWRES=IRES/(W+SAV)\n\nY=F+SQRT(TVEA*TVEA + TVEP*TVEP*F*F)*EPS(1)\n\n$THETA ( 0.0 0.20 1000000.0 ) ; TVCL\n$THETA ( 559 559 559 ) ; TKON\n$THETA ( 0.0 0.25 1000000.0 ) ; TKIN\n$THETA ( 0.0 2.00 1000000.0 ) ; TKDE\n$THETA ( 0.0 1.00 1000000.0 ) ; TVQ\n$THETA ( 2.0 4.00 1000000.0 ) ; TVV1 SART\n$THETA (   2.612   2.612       2.612 ) ; TVV2\n$THETA ( 0.0 0.50 1000000.0 ) ; TVKA\n$THETA (   0.550   0.750       1.000 ) ; TVF\n$THETA ( 0.0 0.30 1000000.0 ) ; TVEP\n$THETA ( 0.0 0.05 1000000.0 ) ; TVEA\n$THETA ( 0.0 0.05  100) ; LAG\n$THETA 1 ; COV1\n\n\n$OMEGA 0.5\n$OMEGA 0 FIX\n$OMEGA 0.5\n$OMEGA 0.5\n$OMEGA 0.5\n$OMEGA 0.5\n$OMEGA 0 FIX\n$OMEGA 0.5\n$OMEGA 0.5\n\n$SIGMA  1 FIX\n\n\n; ================================================\n; PRIOR INFORMATION\n; ================================================\n; THETA Priors -------------------------\n$THETAP\n0.171896943763098     FIXED\n559                   FIXED\n0.123803201342408     FIXED\n1.34635526853289      FIXED\n0.50039445791587      FIXED\n3.22590460107421      FIXED\n2.612                 FIXED\n0.637779367431747     FIXED\n0.590001808531246     FIXED\n0.274684075637132     FIXED\n0.581795606608002     FIXED\n0.0298266165438046    FIXED\n1.56428570804632      FIXED\n; Variance to THETA priors -------------\n$THETAPV BLOCK(13)\n0.00023104719622956   FIXED\n1e-09                 1e-09                \n-1.04034957452488e-05 1e-09                 4.94852473541275e-06 \n-0.000132380576300711 1e-09                 -2.17038124549764e-06 0.0014067543164689   \n3.37273335283364e-05  1e-09                 -2.62117718892531e-05 -9.83509242784927e-05 0.00081696660001029  \n-0.000710872133743716 1e-09                 -5.80489374980969e-05 -0.000732628454663643 -0.000445920587930254 0.0415711705221499   \n1e-09                 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                 1e-09                \n-0.000172472594375445 1e-09                 1.49084134799218e-05  0.000179969508697223  -0.000100076823054098 -0.00224596953903649  1e-09                 0.00218763549223713  \n8.79808997968756e-05  1e-09                 -2.09503076577778e-05 -0.000106886489009519 4.53146822800296e-05  0.00309692392362989   1e-09                 -0.000478502787321484 0.000572988006168684 \n-1.16269709887021e-06 1e-09                 6.45639736622959e-07  -1.04783992387998e-05 -4.49284209190946e-06 8.68593698728809e-06  1e-09                 2.1408518826492e-06   -2.72270313463236e-06 2.26505265649685e-06 \n3.97881008538303e-06  1e-09                 4.80556912014364e-07  2.78111595286765e-05  -3.57590444667545e-06 -4.77538174170191e-05 1e-09                 -6.15715850827622e-05 2.27771577491074e-06  -9.41191843254687e-06 0.000500919762819647 \n5.45053324304558e-07  1e-09                 -1.7273759122248e-07  -1.94999218108576e-07 3.31278452211055e-06  9.25019594253438e-07  1e-09                 -7.15492047000474e-07 5.74018896478916e-08  -1.6453046982605e-07  5.38469288515014e-06  1.20876103617239e-06 \n0.000507900274801002  1e-09                 -5.23098619372009e-05 -0.00029555530917814  0.000667518750332241  -0.0099825252948127   1e-09                 -0.000951749317422468 0.000240924919937593  -2.13162869496016e-05 3.10211676261024e-05  1.05759826160997e-06  0.00723573763017683  \n; OMEGA Priors -------------------------\n$OMEGAP\n0.279195982109823     FIXED\n0                     FIXED\n0.0580908084573451    FIXED\n0.170534475475969     FIXED\n0.0691065862986685    FIXED\n0.0990369682744904    FIXED\n0                     FIXED\n0.464462143306044     FIXED\n0.289551101707683     FIXED\n; Degrees of freedom of OMEGA Prior ----\n$OMEGAPD\n27                    FIXED ; ETCL\n0                     FIXED ; EKON\n316                   FIXED ; EKIN\n216                   FIXED ; EKDE\n11                    FIXED ; ETQ\n71                    FIXED ; ETV1\n0                     FIXED ; ETV2\n176                   FIXED ; ETKA\n33                    FIXED ; ETF\n; SIGMA Priors -------------------------\n$SIGMAP\n1                     FIXED\n; Degrees of freedom of SIGMA Prior ----\n$SIGMAPD\n0                     FIXED ; EPS1\n; ======================================\n$COVARIANCE PRINT=E MATRIX=S\n\n$ESTIMATION METHOD=SAEM INTERACTION NBURN=400 NITER=1000 PRINT=50 NOABORT CTYPE=3\n$ESTIMATION METHOD=IMP NITER=10 PRINT=1 ISAMPLE=3000 EONLY=1\n$TAB ID TIME DV IPRED PRED CMT IRES CWRES IWRES NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=sdtab\n$TAB ID TIME CL KON KSS KINT KDEG Q V1 V2 KA ALAG1 F1 ETAS(1:LAST) NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=patab\n$TAB ID TIME AGE WT BMI CLCR GFR BSLDLC TBSPCSK FBSPCSK NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=cotab\n$TAB ID TIME ISC DISST SEX FIBR EZE STATIN ARM ABDO THIGH STUD NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=catab\n"