Generate a NONMEM control stream for simulations from a run

Generates a NONMEM control stream for simulations from a source run.

nm_simulation(run, n_simulations = 100, seed = sample.int(1e+09, 1),
  output_table_format = ",1PE11.4", output_file = NULL)

Arguments

run
pmxploit NONMEM run object of the source run.

n_simulations
integer. Number of simulations ($SIMULATION NSUBPROBLEMS) to run.

seed
integer. Seed number for random number generation.

output_table_format
(optional) character. NONMEM format (see NONMEM help) for the simulated output table. Default is ,1PE11.4, corresponding to a csv format.

output_file
(optional) character. File path to write the generated control stream to. Default is NULL, not saving any file.

Value

A character string of the generated control stream code.

Details

All THETA, OMEGA and SIGMA parameters will be fixed to the source run final estimates. $ESTIMATION record(s) will be replaced by a single $SIMULATION record (with the specified number of simulations). Simulated output will be saved to a single simtab file ($TABLE).

Examples

# Control stream generation from pmxploit::EXAMPLERUN
cs1 <- nm_simulation(EXAMPLERUN, n_simulations = 300)

# see the output
writeLines(cs1)

#> ; Control stream generated for simulations (N = 300) from:
#> ; ~/pmxploit/example.tar.gz
#> 
#> $PROB  TMDD Qss model for alirocumab
#> ; Reference:
#> ; Djebli N, Martinez JM, Lohan L, Khier S, Brunet A, Hurbin F, et al.
#> ; Target-mediated drug disposition population pharmacokinetics model of alirocumab in
#> ; healthy volunteers and patients: pooled analysis of randomized phase I/II/III studies.
#> ; Clin Pharmacokinet. 2017
#> 
#> $INPUT ID TIME TAD AMT RATE DV MDV EVID CMT ISC DISST AGE SEX WT BMI CLCR GFR BSLDLC FIBR EZE STATIN TBSPCSK FBSPCSK ARM ABDO THIGH STUD
#> 
#> $DATA dataset.csv IGNORE=@
#> 
#> $SUBROUTINE ADVAN13 TOL=5
#> 
#> $MODEL NCOMP=4
#> COMP=(DEPOT)
#> COMP=(SARTOT)
#> COMP=(TPCSK9)
#> COMP=(PERIPH)
#> 
#> $PK
#> 
#> TVCL = THETA(1)
#> TKON = THETA(2)
#> TKIN = THETA(3)
#> TKDE = THETA(4)
#> TVQ  = THETA(5)
#> TVV1 = THETA(6)
#> TVV2 = THETA(7)
#> TVKA = THETA(8)
#> TVF  = THETA(9)
#> TVEP = THETA(10)
#> TVEA = THETA(11)
#> TVLAG = THETA(12)
#> COV1   = THETA(13)
#> 
#> 
#> ETCL   = ETA(1)
#> EKON   = ETA(2)
#> EKIN   = ETA(3)
#> EKDE   = ETA(4)
#> ETQ    = ETA(5)
#> ETV1   = ETA(6)
#> ETV2   = ETA(7)
#> ETKA   = ETA(8)
#> ETF    = ETA(9)
#> 
#> 
#> 
#> CL   = TVCL * EXP(ETCL)
#> KON  = TKON * EXP(EKON)
#> KINT = TKIN * EXP(EKIN)
#> KDEG = TKDE * EXP(EKDE)
#> Q    = TVQ  * EXP(ETQ)
#> 
#> 
#> 
#> COV1V1 = TVV1 * COV1 ** DISST
#> 
#> V1 = COV1V1 * EXP(ETV1)
#> 
#> V2   = TVV2 * EXP(ETV2)
#> KA   = TVKA * EXP(ETKA)
#> ALAG1 = TVLAG
#> 
#> IF (ISC .EQ.0) THEN
#> 
#> F1=1
#> ELSE
#> PHI = LOG(TVF/(1-TVF))
#> F1 = EXP(PHI+ETF)/(1+EXP(PHI+ETF))
#> ENDIF
#> 
#> KEL = CL / V1
#> KPT = Q / V1
#> KTP = Q / V2
#> 
#> KSS = 0.58 + KINT/KON
#> 
#> SARBASE = 0
#> KSYN = TBSPCSK * KDEG
#> 
#> A_0(2) = SARBASE
#> A_0(3) = TBSPCSK * V1
#> 
#> S2=V1
#> S3=V1
#> 
#> 
#> 
#> $DES
#> C2=A(2)/V1
#> C3=A(3)/V1
#> C4=A(4)/V2
#> DADT(1)= -KA * A(1) * ISC * F1
#> 
#> K1 = C2 - C3 - KSS
#> SARF = 0.5*(K1+ SQRT(K1*K1 + 4*KSS*C2))
#> 
#> DADT(2)= KA*A(1)*ISC*F1-(KEL+KPT)*SARF*V1-KINT*A(3)*SARF/(KSS+SARF)+KTP*C4*V2
#> 
#> DADT(3)= KSYN  * V1 - KDEG * A(3)- (KINT - KDEG) * SARF * A(3)/(KSS + SARF)
#> 
#> DADT(4)= KPT* SARF * V1 - KTP * A(4)
#> 
#> ;A(1) = DOSE
#> ;SARF = [SAR_LIBRE]
#> ;A(2) = [SARtot]*V1
#> ;A(3) = [PCSK9tot]*V1
#> ;A(4) = [L tissulaire]*V2
#> 
#> ; PCSK9SAR = TPCSK9 * SARF/(KSS + SARF) => surprediction possible du complexe au
#> ;tps initial !!!
#> 
#> $ERROR
#> 
#> IPRED=F
#> SAV=0
#> IF (IPRED.EQ.0) SAV=0.00001
#> W=F
#> IRES=DV-IPRED
#> IWRES=IRES/(W+SAV)
#> 
#> Y=F+SQRT(TVEA*TVEA + TVEP*TVEP*F*F)*EPS(1)
#> 
#> ; ================================================
#> ; POPULATION PARAMETERS
#> ; ================================================
#> $THETA
#> 0.171896943763098     FIXED ; TVCL
#> 559                   FIXED ; TKON
#> 0.123803201342408     FIXED ; TKIN
#> 1.34635526853289      FIXED ; TKDE
#> 0.50039445791587      FIXED ; TVQ
#> 3.22590460107421      FIXED ; TVV1
#> 2.612                 FIXED ; TVV2
#> 0.637779367431747     FIXED ; TVKA
#> 0.590001808531246     FIXED ; TVF
#> 0.274684075637132     FIXED ; TVEP
#> 0.581795606608002     FIXED ; TVEA
#> 0.0298266165438046    FIXED ; TVLAG
#> 1.56428570804632      FIXED ; COV1
#> $OMEGA
#> 0.279195982109823     FIXED ; ETCL:ETCL
#> 0                     FIXED ; EKON:EKON
#> 0.0580908084573451    FIXED ; EKIN:EKIN
#> 0.170534475475969     FIXED ; EKDE:EKDE
#> 0.0691065862986685    FIXED ; ETQ:ETQ
#> 0.0990369682744904    FIXED ; ETV1:ETV1
#> 0                     FIXED ; ETV2:ETV2
#> 0.464462143306044     FIXED ; ETKA:ETKA
#> 0.289551101707683     FIXED ; ETF:ETF
#> $SIGMA
#> 1                     FIXED ; EPS1:EPS1
#> ; ================================================
#> $SIMULATION (80750138) ONLYSIMULATION NSUBPROBLEMS=300
#> $TABLE ID TIME CMT EVID DV MDV NOHEADER NOPRINT NOAPPEND FORMAT=,1PE11.4 FILE=simtab