Parse a NONMEM control stream

Reads the content of a control stream and extracts informations related to input data, model, estimation and output.

parse_nm_control_stream(filepath = NULL, content = NULL,
  read_initial_values = TRUE, verbose = FALSE)

Arguments

filepath
character. If content = NULL, defines the filepath of the control stream file.

content
character. Text content of a control stream file.

read_initial_values
logical. If TRUE (default), parses initial parameter values from the control stream.

verbose
logical. If TRUE: prints output messages.

Value

A list representing a NONMEM control stream.

Object structure:

content: control stream lines.

problem: problem description.

ignore: list with:

C: logical. TRUE if IGNORE=C

@: logical. TRUE if IGNORE=@

#: logical. TRUE if IGNORE=#

data: data frame containing other IGNORE options.

ignore_at: logical, indicating if if IGNORE=@ is specified.

dataset_file: dataset file name.

subroutine: NONMEM ADVAN subroutine.

model_compartments: tibble of the compartments (name and number of each).

subpopulations: number of subpopulations (for mixture models).

input: tibble of dataset input columns (name, is dropped?, has a synonymn/alias?).

parameters_definitions: tibble of the model parameters (THETAs and ETAs).

estimations: successive estimation methods.

tables: output tables with columns definitions.

Examples

# NOT RUN {
parse_nm_control_stream(filepath = "run_control_stream.con")
# }