Continuous covariates correlations summary

Returns a matrix of the selected continuous covariates correlations.

summarize_covariates_correlations(run, covariates = NULL,
  baseline_only = TRUE, correlation_method = NULL, auto_order = TRUE)

Arguments

run
pmxploit NONMEM run object.

covariates
character vector of continous covariates names. Default is NULL, returning all continuous covariates.

baseline_only
logical. Consider only the baseline (= first) values of the subjects. Default is TRUE.

correlation_method
a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated. If type = "scatterplot", "pearson" method will be used.

auto_order
logical. If type = "heatmap", auto-orders the matrix with respect to the distances between values. Default is TRUE.

split_by
character. Categorical to split the summary by.

Value

A matrix.

Examples

EXAMPLERUN %>% summarize_covariates_correlations()

#>                                     Baseline free PCSK9 (nM)
#> Baseline free PCSK9 (nM)                          1.00000000
#> Baseline total PCSK9 (nM)                         0.63474111
#> Creatinine clearance (mL/min)                    -0.03390245
#> Age (y)                                           0.11602376
#> Glomerular filtration rate (mL/min)              -0.13572416
#> Baseline LDLC (mg/dL)                            -0.05890704
#> BMI (kg/m²)                                       0.14352343
#> Weight (kg)                                       0.10401265
#>                                     Baseline total PCSK9 (nM)
#> Baseline free PCSK9 (nM)                           0.63474111
#> Baseline total PCSK9 (nM)                          1.00000000
#> Creatinine clearance (mL/min)                     -0.06800086
#> Age (y)                                            0.26465204
#> Glomerular filtration rate (mL/min)               -0.14715975
#> Baseline LDLC (mg/dL)                              0.12308252
#> BMI (kg/m²)                                        0.17724420
#> Weight (kg)                                        0.08670552
#>                                     Creatinine clearance (mL/min)     Age (y)
#> Baseline free PCSK9 (nM)                              -0.03390245  0.11602376
#> Baseline total PCSK9 (nM)                             -0.06800086  0.26465204
#> Creatinine clearance (mL/min)                          1.00000000 -0.52845720
#> Age (y)                                               -0.52845720  1.00000000
#> Glomerular filtration rate (mL/min)                    0.67538184 -0.51708487
#> Baseline LDLC (mg/dL)                                 -0.12739227  0.01378944
#> BMI (kg/m²)                                            0.40805133  0.20298742
#> Weight (kg)                                            0.56597426  0.06328087
#>                                     Glomerular filtration rate (mL/min)
#> Baseline free PCSK9 (nM)                                    -0.13572416
#> Baseline total PCSK9 (nM)                                   -0.14715975
#> Creatinine clearance (mL/min)                                0.67538184
#> Age (y)                                                     -0.51708487
#> Glomerular filtration rate (mL/min)                          1.00000000
#> Baseline LDLC (mg/dL)                                       -0.03312037
#> BMI (kg/m²)                                                 -0.15911903
#> Weight (kg)                                                 -0.10956357
#>                                     Baseline LDLC (mg/dL) BMI (kg/m²)
#> Baseline free PCSK9 (nM)                      -0.05890704   0.1435234
#> Baseline total PCSK9 (nM)                      0.12308252   0.1772442
#> Creatinine clearance (mL/min)                 -0.12739227   0.4080513
#> Age (y)                                        0.01378944   0.2029874
#> Glomerular filtration rate (mL/min)           -0.03312037  -0.1591190
#> Baseline LDLC (mg/dL)                          1.00000000  -0.1510291
#> BMI (kg/m²)                                   -0.15102914   1.0000000
#> Weight (kg)                                   -0.16168067   0.7920918
#>                                     Weight (kg)
#> Baseline free PCSK9 (nM)             0.10401265
#> Baseline total PCSK9 (nM)            0.08670552
#> Creatinine clearance (mL/min)        0.56597426
#> Age (y)                              0.06328087
#> Glomerular filtration rate (mL/min) -0.10956357
#> Baseline LDLC (mg/dL)               -0.16168067
#> BMI (kg/m²)                          0.79209179
#> Weight (kg)                          1.00000000

EXAMPLERUN %>% summarize_covariates_correlations(auto_order = FALSE)

#>                                         Age (y) BMI (kg/m²)
#> Age (y)                              1.00000000   0.2029874
#> BMI (kg/m²)                          0.20298742   1.0000000
#> Baseline LDLC (mg/dL)                0.01378944  -0.1510291
#> Creatinine clearance (mL/min)       -0.52845720   0.4080513
#> Baseline free PCSK9 (nM)             0.11602376   0.1435234
#> Glomerular filtration rate (mL/min) -0.51708487  -0.1591190
#> Baseline total PCSK9 (nM)            0.26465204   0.1772442
#> Weight (kg)                          0.06328087   0.7920918
#>                                     Baseline LDLC (mg/dL)
#> Age (y)                                        0.01378944
#> BMI (kg/m²)                                   -0.15102914
#> Baseline LDLC (mg/dL)                          1.00000000
#> Creatinine clearance (mL/min)                 -0.12739227
#> Baseline free PCSK9 (nM)                      -0.05890704
#> Glomerular filtration rate (mL/min)           -0.03312037
#> Baseline total PCSK9 (nM)                      0.12308252
#> Weight (kg)                                   -0.16168067
#>                                     Creatinine clearance (mL/min)
#> Age (y)                                               -0.52845720
#> BMI (kg/m²)                                            0.40805133
#> Baseline LDLC (mg/dL)                                 -0.12739227
#> Creatinine clearance (mL/min)                          1.00000000
#> Baseline free PCSK9 (nM)                              -0.03390245
#> Glomerular filtration rate (mL/min)                    0.67538184
#> Baseline total PCSK9 (nM)                             -0.06800086
#> Weight (kg)                                            0.56597426
#>                                     Baseline free PCSK9 (nM)
#> Age (y)                                           0.11602376
#> BMI (kg/m²)                                       0.14352343
#> Baseline LDLC (mg/dL)                            -0.05890704
#> Creatinine clearance (mL/min)                    -0.03390245
#> Baseline free PCSK9 (nM)                          1.00000000
#> Glomerular filtration rate (mL/min)              -0.13572416
#> Baseline total PCSK9 (nM)                         0.63474111
#> Weight (kg)                                       0.10401265
#>                                     Glomerular filtration rate (mL/min)
#> Age (y)                                                     -0.51708487
#> BMI (kg/m²)                                                 -0.15911903
#> Baseline LDLC (mg/dL)                                       -0.03312037
#> Creatinine clearance (mL/min)                                0.67538184
#> Baseline free PCSK9 (nM)                                    -0.13572416
#> Glomerular filtration rate (mL/min)                          1.00000000
#> Baseline total PCSK9 (nM)                                   -0.14715975
#> Weight (kg)                                                 -0.10956357
#>                                     Baseline total PCSK9 (nM) Weight (kg)
#> Age (y)                                            0.26465204  0.06328087
#> BMI (kg/m²)                                        0.17724420  0.79209179
#> Baseline LDLC (mg/dL)                              0.12308252 -0.16168067
#> Creatinine clearance (mL/min)                     -0.06800086  0.56597426
#> Baseline free PCSK9 (nM)                           0.63474111  0.10401265
#> Glomerular filtration rate (mL/min)               -0.14715975 -0.10956357
#> Baseline total PCSK9 (nM)                          1.00000000  0.08670552
#> Weight (kg)                                        0.08670552  1.00000000