Parameters correlations

Returns the selected parameters correlations in the form of either a matrix, a data frame or a plot.

summarize_parameters_correlations(run, parameters = NULL,
  baseline_only = TRUE, correlation_method = NULL, auto_order = TRUE)

Arguments

run
pmxploit NONMEM run object.

parameters
character vector of parameters names. Default is NULL, returning all individual parameters (random and post-hoc).

baseline_only
logical. Consider only the baseline (= first) values of the subjects. Default is TRUE.

correlation_method
a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated. If type = "scatterplot", "pearson" method will be used.

auto_order
logical. If type = "heatmap", auto-orders the matrix with respect to the distances between values. Default is TRUE.

Value

A matrix.

Examples

EXAMPLERUN %>% summarize_parameters_correlations()

#> Correlations are not computed for parameters(s) with one unique value: KON, V2, ALAG1

#>               V1          F1           Q       KSS      KINT          KA
#> V1    1.00000000 -0.58741001 -0.05479555 0.1106159 0.1098006  0.30323370
#> F1   -0.58741001  1.00000000  0.13141825 0.2147780 0.2165945 -0.06226802
#> Q    -0.05479555  0.13141825  1.00000000 0.3817025 0.3792538  0.05201759
#> KSS   0.11061590  0.21477798  0.38170253 1.0000000 0.9981868  0.19737220
#> KINT  0.10980058  0.21659448  0.37925379 0.9981868 1.0000000  0.19826681
#> KA    0.30323370 -0.06226802  0.05201759 0.1973722 0.1982668  1.00000000
#> CL    0.30095621  0.02300692  0.09799981 0.3480292 0.3480398  0.22345707
#> KDEG  0.16623895 -0.04360299  0.09469923 0.2157855 0.2162849  0.17495309
#>              CL        KDEG
#> V1   0.30095621  0.16623895
#> F1   0.02300692 -0.04360299
#> Q    0.09799981  0.09469923
#> KSS  0.34802922  0.21578546
#> KINT 0.34803981  0.21628493
#> KA   0.22345707  0.17495309
#> CL   1.00000000  0.33390241
#> KDEG 0.33390241  1.00000000

EXAMPLERUN %>% summarize_parameters_correlations(auto_order = FALSE)

#> Correlations are not computed for parameters(s) with one unique value: KON, V2, ALAG1

#>              CL       KSS      KINT        KDEG           Q          V1
#> CL   1.00000000 0.3480292 0.3480398  0.33390241  0.09799981  0.30095621
#> KSS  0.34802922 1.0000000 0.9981868  0.21578546  0.38170253  0.11061590
#> KINT 0.34803981 0.9981868 1.0000000  0.21628493  0.37925379  0.10980058
#> KDEG 0.33390241 0.2157855 0.2162849  1.00000000  0.09469923  0.16623895
#> Q    0.09799981 0.3817025 0.3792538  0.09469923  1.00000000 -0.05479555
#> V1   0.30095621 0.1106159 0.1098006  0.16623895 -0.05479555  1.00000000
#> KA   0.22345707 0.1973722 0.1982668  0.17495309  0.05201759  0.30323370
#> F1   0.02300692 0.2147780 0.2165945 -0.04360299  0.13141825 -0.58741001
#>               KA          F1
#> CL    0.22345707  0.02300692
#> KSS   0.19737220  0.21477798
#> KINT  0.19826681  0.21659448
#> KDEG  0.17495309 -0.04360299
#> Q     0.05201759  0.13141825
#> V1    0.30323370 -0.58741001
#> KA    1.00000000 -0.06226802
#> F1   -0.06226802  1.00000000