Summarize shrinkage

Summarize shrinkage (ETA, EBV and EPS) of the individual parameters resulting of an estimation step.

summarize_shrinkage(run, estimation_number = NULL, type = NULL)

Arguments

run
pmxploit NONMEM run object.

estimation_number
integer. Number of the estimation step. Default is NULL, returning the last estimation step.

type
character vector. Shrinkage type (ETA, EBV or EPS).

Value

A data frame with one row per shrinkage value and the following columns:

type: shrinkage type (ETA, EBV or EPS)

parameter: random parameter name

shrinkage: shrinkage value

Examples

EXAMPLERUN %>% summarize_shrinkage()

#> # A tibble: 19 x 3
#>    type  parameter shrinkage
#>    <chr> <chr>         <dbl>
#>  1 ETA   ETCL         0.626 
#>  2 ETA   EKON         1     
#>  3 ETA   EKIN         0.204 
#>  4 ETA   EKDE         0.265 
#>  5 ETA   ETQ          0.755 
#>  6 ETA   ETV1         0.327 
#>  7 ETA   ETV2         1     
#>  8 ETA   ETKA         0.477 
#>  9 ETA   ETF          0.438 
#> 10 EBV   ETCL         0.644 
#> 11 EBV   EKON         0     
#> 12 EBV   EKIN         0.200 
#> 13 EBV   EKDE         0.273 
#> 14 EBV   ETQ          0.765 
#> 15 EBV   ETV1         0.322 
#> 16 EBV   ETV2         0     
#> 17 EBV   ETKA         0.492 
#> 18 EBV   ETF          0.443 
#> 19 EPS   EPS1         0.0626

EXAMPLERUN %>% summarize_shrinkage(type = "ETA")

#> # A tibble: 9 x 3
#>   type  parameter shrinkage
#>   <chr> <chr>         <dbl>
#> 1 ETA   ETCL          0.626
#> 2 ETA   EKON          1    
#> 3 ETA   EKIN          0.204
#> 4 ETA   EKDE          0.265
#> 5 ETA   ETQ           0.755
#> 6 ETA   ETV1          0.327
#> 7 ETA   ETV2          1    
#> 8 ETA   ETKA          0.477
#> 9 ETA   ETF           0.438